Added Multipart support for ASP.NET hosting
This is dependent on the ASP.NET hosting environment and uses the
HttpRequestContextBase to access both Form and File data.  The Nancy Request
class has also been extended to allow callbacks to be used to populate both
Form and File data, which will provide future extensibility points for use with
other implementaitons.